Fix bug in PeekWrappedP
It worked correctly when inbox was drained using peek+remove. If it was
drained using `poll()` there was an `NPE` and secondly, `poll()` returned
`true` instead of the polled object.

Second improvement is that when using peek+remove, the item is logged in
`peek`, not in `remove`, which is called after processing the item in
`AbstractProcessor`, which in turn caused the output items being logged before
input items in case of `peekInput(peekOutput(processor))`.

Another bug was not using `Distributed` version of function.

Also updated the test to test these scenarios.