Add fair distribution of tasklets
Previously, tasklets for a job were assigned to cooperative workers always
starting from worker 0. This could lead to the last worker to have less
tasklets assigned, or even none in extreme cases (such as many small DAGs with
local parallelism of 1 on all vertices, a setting typical if there are many
jobs).

This change makes the starting index global. Work stealing didn't help here
because it only takes place when worker dismissed some tasklet, which works
with assumption that it was balanced before.