Implement getJobs() api
There are 2 types of jobs: Submitted job and Tracked job
The client / jet instance which submits a job has a submitted job object. Upon master failures, it re-submits the job to the next master. It is already the current approach. Other jet instances can get a proxy for this job, using tracked job objects. They can listen for result of the job, query its status or cancel it.