Fix CIES to not reorder WM and barrier
It could happen that even in exactly once mode, the WM and barriers
could be reordered with respect to each other.

This was caused by the fact that we could stop at a barrier in one queue,
that would cause the barrier to be emitted, but we continued to another
queue which drained a watermark, which would also cause a WM to be
emitted. After iterating all queues, we first checked for WM to emit and
then for barrier to emit, effectively causing the reordering. It was
fixed by moving the emit check inside to the loop and breaking the
iteration if we emitted either WM or barrier.

ItemDetector was simplified too along the way.