Add automatic WM throttling to Pipeline API
Also modify WatermarkEmissionPolicy: it will tell what the new watermark
should be, not just whether to forward or not. For example, if
frameSize=1000 and new WM from the policy is 1234, it will not just tell
to forward 1234, but it will tell to forward 1000. This way WMs from all
parallel source instances will align and the WMs will be really
throttled.

It's implemented in `Planner.createdDag`: it checks frameDefs from all
transforms, calculates GCD and then changes emission policies in all WM
generating transforms.

Fixes  #743