Don't shrink rwin when during lulls
We gradually expand and shrink rwin based on the amount of data we've processed during the flow control period. This works well if the stream of data is regular. If it is burst-and-lull (i.e. a burst of data and then nothing for a while), then during the lull the rwin shrinks to 1. When the burst occurs, we apply a lot of backpressure unnecessarily.

This fix works around the problem by not shrinking the rwin if there are no items waiting in the inbox. The `call` method processes data from inbox. During a lull, the inbox is normally empty - we avoid shrinking the rwin in this situation. When inbox is not-empty, it means that we are actively processing data and shrinking/expanding works as before. Expansion of the rwin is unaffected by this change.

Fixex #785