Fix race when snapshot started after job was done
Fixes this scenario: the job completes (normally or abruptly). It
deletes snapshot maps and then it marks CompletionFuture in
MasterContext as done.

In parallel, scheduled snapshot is about to start by calling
`JobCoordinationService.beginSnapshot`. It checks whether the
MasterContext is done. It sees it is not, so it goes to generate
snapshotId by looking at `LATEST_STARTED_SNAPSHOT_ID_KEY`, but it's null
since the map was deleted in parallel. So it starts the snapshot 0.

The `SnapshotOperation` will be rejected on members, but it recreates
the maps, which causes the maps to leak and the test to fail.

The fix reorders operations in `MasterContext.finalizeJob`: it first
sets the MasterContext.jobStatus, completes the future and then goes on
to delete maps.

Fixes #865