Prefer using job name in logs
The following is used in log messages:

`AbstractJobProxy`: logs on the submitting client or member
  xxxx-xxxx-xxxx-xxxx (name ??)   - if config not loaded yet
  xxxx-xxxx-xxxx-xxxx (name '')   - if loaded, but name null
  xxxx-xxxx-xxxx-xxxx (name 'abc')

`MasterContext`: logs on coordinator member. Name is always known.
  job 'abc', execution xxx-xxx-xxx-xxx                - if name set
  job 'xxx-xxx-xxx-xxx', execution xxx-xxx-xxx-xxx    - if name not set

`ExecutionContext`: on each member, name not known for a short while
initially
  job 'abc', execution xxx-xxx-xxx-xxx              - if name set
  job xxx-xxx-xxx-xxx, execution xxx-xxx-xxx-xxx    - if name not set

Initial log, printed on each member on INFO level to give association for
jobId, jobName and executionId:

  Execution plan for jobId=xxx-xxx-xxx-xxx, jobName=null/'abc',
  executionId=xxx-xxx-xxx-xxx initialized

The important change is that if job name is set, job ID is not always
displayed. If the job name is not unique, we won't be able to tell the
jobs apart in logs. This is documented at `JobConfig.setName()` method.

Fixes #911