Fix error handling when deleting snapshots
Fixes failure-2 in #1009

This happened:
- graceful restart was requested. Terminal snapshot 1 was done, execution
completed
- before MC.jobStatus was changed to NOT_RUNNING, another snapshot 2 was
started based on scheduled interval. It failed on members, because the
execution was done. This is ok.
- because the snapshot 2 failed, it was deleted in
`onSnapshotCompleted`. In parallel, job was being restarted where all
except the current successful snapshot are deleted. But the deletion
failed with NPE in entry processor, because the SnapshotRecord was
already deleted. Exception wasn't caught and interrupted the
`MC.tryStartJob`, leaving the job in STARTING state.

The fix is to not cause the exception when deleting snapshots and ignore,
if any other should happen.