Fix WmCoalescingTest failure
It's hard to tell for sure what happened here, because the output of all
parallel tests in the class is mixed, but looks like the map-writing
thread was stalled for >2 seconds and the source marked partition as
idle. This should not happen if there is output all the time.

In the fix, I increased the idle timeout to 4 seconds and made the
updating thread to run until interrupted. I also made the value put to
map different in each test so that we can tell apart output of each test
and prove the cause should it happen again.

Fixes #1018