Make single SnapshotRestore input higher priority
Snapshot is not started until all higher-priority edges are completed.
Snapshot restoring is done using such edges. However, in the case of
source, the restore edge wasn't higher priority, because it's the only
edge. It can happen that the first snapshot is done before the state of
the source is restored.

Fixes #1100