Reuse snapshot maps
Changes the mode of operation where we created new maps for each
snapshot and destroyed the maps when deleting the snapshot to a mode
where we have 2 maps and use one for writing and one for storing.

Notes:
- responsibility for updating JobRecord moved from JobCoordinatinService
to MasterContext

- MasterContext contains final field with JobRecord. It mutates it
(therefore some fields in JobRecord were made volatile) and writes it
out after each change using an async call. Since it can be mutated in
parallel and parallel async writes can be reordered, we use timestamp to
ensure that an older update doesn't overwrite a newer one.

- One snapshot is stored in one map (previously, each vertex used its
own map). The chunk key now contains vertex name and snapshotId.
SnapshotId was added to it to distinguish zombie writes, should they
happen, or writing after a failed `IMap.clear`. When reading, unknown
vertices and snapshotIds are logged and ignored.

- Which data map is used for new snapshot is communicated to members
using SnapshotOp. Because target member must know this before it starts
writing, we've removed the (rare) possibility that snapshot is written
before the operation is received.