Simplify saveToSnapshot contract
The contract of `saveToSnapshot` required to first offer the possibly
half-emitted item to the outbox and only then proceed to emitting to
snapshot. This was a major complication and a gotcha. Also, in case of
batch sources such as `readFileP` if, for whatever reason snapshots are
enabled, it could happen that the outbox rejected the snapshot barrier
offered by the ProcessorTasklet due to an unfinished half-emitted item
(or worse, if assertions were disabled and some output queues won't have
the barrier).

This commit fixes it by adding `OutboxImpl.block()` that would only
allow to offer the unfinished item to the outbox and `ProcessorTasklet`
calls `complete()` until there's no unfinished item in the outbox. This
allows us to remove the requirement from `saveToSnapshot`'s contract.