Add run-until-output-matches mode to TestSupport
Removes the `disableRunUntilCompleted()` method. The method was meant to
test streaming sources, but the timeout needs to be pretty high to
account for test environment hiccups which makes the test slow even in
cases when the output is produced very shortly after run. This commit
introduces `runUntilOutputMatches()` which runs `complete()` until the
output matches and for little extra time to check that there's no more
output.

Fixes #1261