Fix migration taking place before test job starts
In the test we submitted the job and immediately proceeded to add a
member to cause migration, there was no synchronization. It could happen
that the member was added before the job started, which caused the test
to fail.

This commit also uses random group name for the remote members so that
they don't inadvertently form a cluster with test running in another JVM
on the same network.

Fixes #1108