Fix race in shutting down of cooperative workers
This scenario was possible:

- StartExecutionOp is received. `JobExecutionService.beginExecution()`
  is entered. Member is not yet shut down, so the `isShutdown` check
  passes.

- member is now shut down gracefully (in another thread). First 2 lines
  of this code are executed:
```
        jobCoordinationService.shutdown();
        jobExecutionService.shutdown(true);
        taskletExecutionService.shutdown(true);
```

- `JobExecutionService.beginExecution()` continues and calls
  `ExecutionContext.beginExecution`, which calls
  `taskletExecService.beginExecute()`

- in `TaskletExecutionService.beginExecution` the shutdown check passes

- now the 3rd line from the code above is executed:
  `taskletExecutionService.shutdown(true);`

- now `TaskletExecutionService` is shut down, so CooperativeWorkers can
  terminate, but we proceed to add tasklet to a dead worker and that
  tasklet won't get completed. Master will remain stuck waiting for all
  tasklets to complete.

Resolution is like this: we don't stop cooperative worker on graceful
shutdown, only on forceful shutdown. The process of graceful shutdown is
coordinated by the master, the master terminates all executions and knows
when they terminated. When exeecutions are down, master notifies the
member and it continues with forceful shutdown. There's no need to
synchronize shutdown and tasklet addition in `TaskletExecutionService`,
because now there's only forceful shutdown and we don't care if tasklets
are added, we're going down anyway.

Fixes #918, fixes #1156