Fix StreamKafkaP_TimestampModesTest failure
The issue was that the kafka source total parallelism was 4, but there
was only 1 partition in the test. So 3 processors were idle. Most of the
time, the idle processors immediately emitted IDLE_MESSAGE. But in the
failed run, IDLE_MESSAGE was emitted after the active processor emitted
the items and watermarks. This caused that wm(1) was suppressed at the
coalescer and only wm(2) was forwarded after the idle message was
received.

It's resolved by using total parallelism=1 everywhere, the test isn't
influenced by the parallelism.

Fixes #1294