Test `jet.sh submit` and fix options for the jar
The syntax of submit command is

  jet.sh submit [options] <jarFile> [<argsForJarFile>...]

However, if the argument for jar file contained an option starting with
"-", it was an option of the submit command and the script failed. This
was fixed by setting a flag on picocli parser.

Also adds test for the `submit` command. In `JetBootstrap` a possibility
to set the client was added.

Fixes #1401