Always initialize executionStartTime
Fixes #1436.

It fixes the execution duration when the `executionStartTime` was not
initialized and was 0. On Linux `System.nanoTime` returns 0 at about OS
boot, which explains the 8 hours of runtime. The value was printed
because we've set the `executionStartTime` deeply in `tryStartJob`, but
this time the job was finalized before `tryStartJob` was called. Also
moves the initialization of `executionStartTime` at the very beginning
of `tryStartJob` because other exceptions could lead to garbabe duration
logged.

The commit does not fix the "execution 0000-0000-0000-0000" - it means
that no execution began yet. It might be surprising, but it's at least
not a garbage value.