Fix race in JobCoordinationService.terminateJob()
It was possible that the method saw neither the JobResult nor the
MasterContext because between those two checks another thread created
the JobResult and removed the MasterContext. In this case it threw
`JobNotFoundException`, which was the failure in the issue.

Now we first check `masterContexts` and if one isn't found, we proceed
to checking the JobResults. When finalizing the job, the creation of
JobResult _happens before_ removal of MasterContext.

Fixes #1519