Drop unstable mapStateful_keyedWithTtl, add coverage
The test was unstable because when processor has total parallelism > 1,
the watermarks are non-deterministic. They can be postponed if waiting
for another processor. The unit test covered the TTL handling enough.

I also added a test for the interrupted cleanup logic in
`tryProcessWatermarks` and more, the unit test coverage is 100% now.

Fixes #1560