Reduce memory usage of JobSubmissionRegressionTest
The test is using a lot of memory when there is a lot of cores:
Each job creates core-count * core-count amount of queues, and each queue
has 1024 elements. On jenkins with 72 cores, this causes about 50mb of
memory usage per job, and the test submits 72 jobs at the same time.-

The fix is to set the queue size on edge to 1 as this is not related
to the regression tested by this test.