Ignore memberAdded for the local member
Fixes #1627

The test failed because the memberAdded was executed while the test was
already running. The memberAdded is an async method, when it gets to
increment the counter, jobs can be already reading maps. This PR fixes
the issue only if there is just 1 member by ignoring memberAdded event
for the local member. For multiple members it can still happen and
there's no easy way to fix it.

The issue is very minor. It's very unlikely that you'll start a batch
map-reading job in a cluster that you just started. You'd at least
insert some entries into the map and while doing that the memberAdded
handler will have a chance to increment the counter. The test that
failed was testing the reading of an empty map, and even there it failed
for the first time.