Don't assert unfinished item after completeEdge
We normally allow unfinished items after `process()` even when the inbox is
empty. We proceed to `saveToSnapshot()`, `tryProcess()`, `completeEdge()` or
`complete()` as needed and those should deal with it. This is only an issue
with async processors, sync processors finish emitting the items before they
remove the item from the inbox.

We could do a similar process as with do with `complete()`: if `complete()`
returns with an unfinished item in the outbox, we block the outbox and call
`complete()` until it's finished, before proceeding to other methods (such as
to `saveToSnapshot()`). However, doing similar thing with the `process()`
method makes no sense since it's normally implemented to only act on items in
the inbox and the inbox would be empty. We could call `tryProcess()` until
there's no unfinished item, but the async processor again doesn't need it - the
`tryProcess()` method is essentially `process()` method called with an empty
inbox. Merging these two methods might even simplify things.

We could ensure that there's no unfinished item before interrupting the
`process/tryProcess/completeEdge` to proceed to `saveToSnapshot`, but current
impls of the async processors don't need it: it would not be enough for them,
they need to flush more than just the current item anyway.

This PR removes two asserts for unfinished item and takes the advantage from
this in the async processors.

Fixes #1716