Fix race when starting the snapshot
While the lock was held, we decided to proceed with the snapshot. Then we
released the lock and in another thread, the job was restarted. And we started
two snapshots in parallel. The fix is to copy the executionId while the lock is
held, if the job restarts before the snapshot executes, the operations will be
ignored due to a newer executionId.