Enable heap dump on OOM
We put the setting directly to pom as we want to use it in all builds
and 1G heap should be fine to upload to S3.

The hprof file will be put to root directory of the module which runs
the test.

In case there are multiple heap dumps we can link to the file
to particular test via pid.