Add retry logic to elastic client calls
Elastic client tries to connect to a node only once for each request
(meaning a call to the client here, not actuall HTTP request). When a
request fails the node is marked dead and is not retried for the
request.

This causes issues with single node cluster or small clusters where all
nodes become unreacheble for short amount of time).

For different requests
- the dead nodes are retried after 1 m.
- when all nodes are marked dead one of the dead nodes is retried.