Avoid shading already shaded jar
Maven jar plugin recreates the original jar only
if the jar is missing or there are changes in the inputs.
The shaded jar contains all inputs up to date so it is not
changed and is used as input for shade plugin if you run
the build 2nd time. Causing the build to fail due to overlapping
resources.

See https://maven.apache.org/plugins/maven-jar-plugin/jar-mojo.html#forceCreation