Fix client.close() verification in Elastic sink test
The default parallelism for the Sink is 2, so 2 clients are created.
The ClientHolder in the test stored only one of the 2 clients created.
Sometimes the same client instance is returned for both threads,
causing the issue #2427.

This commit changes so all clients are stored in the holder and verified.

Also adds a test for closing the client in a non-failing Sink.

Fixes #2427