Fix JobCoordinationService race
The `JobCoordinationService.getJobStatus()` only checked the
`JobExecutionRecord`, but not the `JobRecord`. However, right after the
`JetInstance.newJob()` returns, only the latter is guaranteed to exist.
It could happen that `getJobStatus()` could throw `JobNotFoundException`
if called right after submitting the job using `newJob`.

The failure was rare because after `newJob` the `MasterContext` always
exists. But in the test we regularly restart cluster members and it
could happen after a new master takes over: the `MasterContext` isn't
created yet, nor is `JobExecutionRecord` or `JobResult`, but we didn't
look at `JobRecord`. With some luck, the issue is reproducible by adding
a 1-second sleep at the beginning of `tryStartJob()`.

The fix is to make `jobRecordHandler` parameter of `callWithJob()`
method mandatory.

The NPE was caused by `toNullFunction` returning a non-null function for
a null consumer. We fix this by making the `consumer` parameter
non-null.

Fixes #2481