Move the LP determination logic to Job Planning Phase
Move the local parallelism determination logic of a Pipeline job to the Job
planning phase. Add this determination logic into the addToDag() method
of the Transform classes. Remove the Vertex.determineLocalParallelism() calls.
To avoid reordering, most transforms had to match the local parallelism of
their upstream transforms. Match the local parallelism of the upstream
transforms for these transforms. This could only be done during the Job
planning phase.
Note: This is a partial commit, parallelism determination are not yet
completed for some transforms that are hard to consider.