Fix job hanging due to lost DONE_ITEM
The failure scenario was this:

- SenderTasklet reaches EOS in tryFillInbox(), writes DONE_ITEM

- We try to send the last packet in the call method, Connection.write()
returns true

- The connection is broken and the packet is never sent. The connection
is transparently reestablished, but the packet isn't sent again.

- Jet doesn't use this connection anymore in SenderTasklet as EOS has
been reached, so the failure goes unnoticed

- Remote member never receives the packet, the ReceiverTasklet will keep
waiting for the DONE_ITEM

The solution is that when the ReceiverTasklet creates data for the flow
control packet, it compares that the Connection object it was crated
with is the same.

We send the flow control packet regularly so any transparent
reconnection will cause the job to fail. Currently it would fail only if
some SenderTasklet tried to send something.

Fixes #2158