Fix AbstractJobProxy issues
- fix when `Job` instance is from non-coordinator and the coordinator
leaves, the job completes. The issue was that in a specific case,
`HazelcastInstanceNotActiveException` was thrown on the
`JoinJobOperation` target member (the master) and the non-coordinator
didn't retry the operation. The fix is to retry the operation in case of
`HazelcastInstanceNotActiveException`, but only if the local member is
running and the exception likely originates on the remote member. Fixes
#2461

- `AbstractJobProxy.resubmitJob()` and `rejoinJob()` checked
`AbstractJobProxy.memberUuid()` for a `null` result, but it never
actually returned null but threw instead. The fix is to specify a single
exception to throw (the `IllegalStateException`) and fix the code to
check for the exception.

- Unify the code of `SubmitJobCallback` and `JoinJobCallback` into a
base class.