Fix determinePreserveOrderFlags Algorithm
Formerly, it would not work in pipelines containing multiple branches.
The problem was determining the flag of the stage where the pipeline
split into two branches and suppose that the flag of one branch is false
and the other's is true. Give priority to "true" value and propagate this
 true value to upstream transforms and fix the issue. Get this by not
propagating false values, anyway the default value of these flags is
false and if there is no downstream transform having true value, it
will remain false.