Reduce poll and init timeout in Kafka source
We reduced the `poll` timeout from 50ms to 0ms. The consumer receives
records on the background, there's no need to block for more records,
just grab those that are immediately available. This reduces snapshot
latency.

Secondly, it fixes the issue when the `consumer.partitionsFor()` can
block for up to 60 seconds waiting for the metadata in the `init()`
method. We reduced the waiting to 1 second, which can easily not be
enough. As as consequence, we had to adapt the processor to live without
partition information and try to get it later (we already did that when
checking for partitions added at runtime). The major change was that
now, when restoring offsets from the snapshot, we can't rely on
`currentAssignment`, but we'll simply assume the topic has at least as
many partitions as the number of offsets we restored for that topic.

As a side effect, this also fixes the issue when a partition was added
while the job was down. Take this scenario:
- topic T has 1 partition
- the job is suspended
- while suspended, another partition is added
- the job wakes up

Now there's the change: Previously, the processor found out in the
`init` method that there are 2 partitions. These partitions would start
at default positions. Then, offset for partition 1 was restored, but no
offset for partition 2. If the `auto.offset.reset` parameter was
`latest` (the default), we could miss the messages in partition 2 before
the job restarted.

With current implementation we don't check partition counts in `init`.
When the job restarts, it will restore offsets for partition 1, but not
for 2. The partition count for the topic will be inferred to 1. Later,
after the snapshot is restored, it will query the metadata for the topic
and find out there are 2 partitions. The 2nd partition will be added and
sought to the beginning and all items from it will be consumed.

Fixes #2724